# Remove date regexp