**Issue:** https://github.com/pelletier/go-toml/issues/440

Hand-roll the date matching logic to avoid trying to match a regexp on
every integer.

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkUnmarshalToml-8     293449        272134        -7.26%

benchmark                    old allocs     new allocs     delta
BenchmarkUnmarshalToml-8     2746           2650           -3.50%

benchmark                    old bytes     new bytes     delta
BenchmarkUnmarshalToml-8     133604        127548        -4.53%
```

A lot of the new code looks duplicated. It's not quite. I found it easier to read and more precise in the reported errors than the various refactors I tried.

@moorereason